- Store the amount of uploaded puzzles of an identity as a property in the identity, this allows your introduction client to know how many puzzles an identity uploads per day. It will use a random index in [0, puzzles_per_identiy_per_day] for downloading puzzles. This allows us to run the seed identity with 1000 puzzles per day for example. The random distribution of downloads will push the probability of downloading a puzzle which someone else has solved already to zero if the puzzle amount is large enough. - Download a maximum of 3 puzzles per identity per day - Use UTC time for puzzles
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/plugins/WoT@23830 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/plugins/WoT/introduction/IntroductionServer.java
@@ -54,7 +54,7 @@ public final class IntroductionServer implements Runnable, ClientCallback {
 	private static final int STARTUP_DELAY = 1 * 60 * 1000;
 	private static final int THREAD_PERIOD = 30 * 60 * 1000; /* FIXME: tweak before release */
 
-	public static final byte PUZZLE_COUNT = 10; 
+	public static final int PUZZLE_COUNT = 10; 
 	public static final byte PUZZLE_INVALID_AFTER_DAYS = 3;
 	public static final int PUZZLE_REINSERT_MAX_AGE = 12 * 60 * 60 * 1000;
